moved more sprite-suppliers
