+++ engine/graphics/grenade-sprite-supplier.go
@@ -1,24 +1,23 @@
-package engine
+package graphics
 
 import (
 	"retro-carnage/assets"
 	"retro-carnage/engine/geometry"
-	"retro-carnage/engine/graphics"
 )
 
 const grenadeImagePath = "images/weapons/grenade.png"
 
 var (
-	grenadeSprite *graphics.SpriteWithOffset
+	grenadeSprite *SpriteWithOffset
 )
 
 // GrenadeSpriteSupplier is used to provide sprites for grenades.
 type GrenadeSpriteSupplier struct{}
 
-func (gss *GrenadeSpriteSupplier) Sprite(int64) *graphics.SpriteWithOffset {
+func (gss *GrenadeSpriteSupplier) Sprite(int64) *SpriteWithOffset {
 	if nil == grenadeSprite {
 		var sprite = assets.SpriteRepository.Get(grenadeImagePath)
-		grenadeSprite = &graphics.SpriteWithOffset{
+		grenadeSprite = &SpriteWithOffset{
 			Offset: geometry.Point{},
 			Source: grenadeImagePath,
 			Sprite: sprite,
